+++ lib/di/di_init.dart
@@ -5,7 +5,7 @@ import 'package:flutter_base_rootstrap/data/data_sources/remote/abstract/auth_da
 import 'package:flutter_base_rootstrap/data/data_sources/remote/abstract/product_data_source.dart';
 import 'package:flutter_base_rootstrap/data/data_sources/remote/concrete/product_data_source_impl.dart';
 import 'package:flutter_base_rootstrap/data/network/config/network_config.dart';
-import 'package:flutter_base_rootstrap/data/repositories/authentication_repository_impl.dart';
+import 'package:flutter_base_rootstrap/data/repositories/auth_repository_impl.dart';
 import 'package:flutter_base_rootstrap/data/repositories/product_repository_impl.dart';
 import 'package:flutter_base_rootstrap/devices/permissions/abstract/permission_manager.dart';
 import 'package:flutter_base_rootstrap/devices/permissions/concrete/mobile/_permissions_android.dart';
